requirements.txt
sanic
python-dotenv
aiohttp
Fabric3
